third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-004-expected.txt


FAIL CSS Test: @viewport constrained - min-width 10000px assert_equals: expected 10000 but got 800
